SampleRobot: Remove runtime warning
Given that WPILib mainly marked SampleRobot as deprecated to give teams
a large warning in their IDEs and at compile time, we have decided that
this is probably a bad idea until we come up with a nicer solution.

This also restores the large orange warning box in the docs.